requirements.txt
requests[security]
fake-useragent
xmltodict
pyjsbn-rsa